# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Node practice",
    "description": "Test project to create simple e-shop backend",
    "contact": {
      "name": "Swagger API Team",
      "email": "apiteam@swagger.io",
      "url": "http://swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "definitions": {
    "User": {
      "properties": {
        "id": {
          "type": "number",
          "format": "double",
          "x-nullable": true
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password",
        "firstName",
        "lastName"
      ],
      "type": "object"
    },
    "Category": {
      "properties": {
        "id": {
          "type": "number",
          "format": "double",
          "x-nullable": true
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "parentCategory": {
          "type": "string"
        },
        "childCategories": {
          "type": "string"
        },
        "products": {
          "type": "string"
        }
      },
      "required": [
        "name"
      ],
      "type": "object"
    },
    "Product": {
      "properties": {
        "id": {
          "type": "number",
          "format": "double",
          "x-nullable": true
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "categories": {
          "type": "string"
        }
      },
      "required": [
        "name"
      ],
      "type": "object"
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get all users",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/User"
              },
              "type": "array"
            }
          }
        },
        "parameters": []
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create new user",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/User"
              },
              "type": "array"
            }
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "new user",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ]
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login as user",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/User"
              },
              "type": "array"
            }
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "user credentials",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ]
      }
    },
    "/users/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/User"
              },
              "type": "array"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get user by id",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/User"
              },
              "type": "object"
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number"
          }
        ]
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update user by id",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/User"
              },
              "type": "array"
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number"
          },
          {
            "in": "body",
            "name": "new user",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete user by id",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/User"
              },
              "type": "array"
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number"
          }
        ]
      }
    },
    "/categories": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "Get all root categories",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/Category"
              },
              "type": "array"
            }
          }
        },
        "parameters": []
      },
      "post": {
        "tags": [
          "categories"
        ],
        "summary": "Create new category",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/Category"
              },
              "type": "array"
            }
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "new user",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "Get category by name",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/Category"
              },
              "type": "object"
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ]
      },
      "put": {
        "tags": [
          "categories"
        ],
        "summary": "Update category by id",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/Category"
              },
              "type": "array"
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number"
          },
          {
            "in": "body",
            "name": "new category",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "categories"
        ],
        "summary": "Delete category by id",
        "produces": [
          "application/json"
        ],
        "responses": {
          "204": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/Category"
              },
              "type": "array"
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number"
          }
        ]
      }
    },
    "/categories/products/{id}": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Get all category products",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/Product"
              },
              "type": "array"
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number"
          }
        ]
      },
      "post": {
        "tags": [
          "products"
        ],
        "summary": "Create new product",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/Product"
              },
              "type": "array"
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number"
          },
          {
            "in": "body",
            "name": "new product",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Get product by id",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/Product"
              },
              "type": "array"
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number"
          }
        ]
      },
      "put": {
        "tags": [
          "products"
        ],
        "summary": "Update product by id",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/Product"
              },
              "type": "array"
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number"
          },
          {
            "in": "body",
            "name": "product updates",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "products"
        ],
        "summary": "Delete product by id",
        "produces": [
          "application/json"
        ],
        "responses": {
          "204": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/Product"
              },
              "type": "array"
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number"
          }
        ]
      }
    },
    "/user/me": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "Get current user",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/User"
              },
              "type": "object"
            }
          }
        },
        "parameters": [
        ]
      },
      "put": {
        "tags": [
          "me"
        ],
        "summary": "Update current user",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "schema": {
              "items": {
                "$ref": "#/definitions/User"
              },
              "type": "object"
            }
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "current user updates",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ]
      }
    }
  },
  "produces": [
    "application/json"
  ]
}